Tests for MultiLogger and ConsoleLogger
