+++ Tests/QTFK.Core.Tests/QTFK.Core.Tests.csproj
@@ -53,7 +53,7 @@
   </Choose>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ConsoleLoggerTests.cs" />
+    <Compile Include="LoggerTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\QTFK.Core\QTFK.Core.csproj">
